# services/billing/api/openapi.yaml
openapi: 3.1.0
info:
  title: Korp Billing API
  version: 1.0.0
  description: API for simplified invoices and durable closure operations.
servers:
  - url: /api/billing/v1
paths:
  /metrics:
    servers:
      - url: /
    get:
      operationId: billingMetrics
      summary: Prometheus metrics for the Billing API and durable closure operations.
      responses:
        '200':
          description: Prometheus text exposition.
          content:
            text/plain:
              schema: { type: string }
  /invoices:
    post:
      operationId: createInvoice
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Invoice created.
          headers:
            Idempotent-Replayed:
              description: Present and true when the original response is replayed.
              schema: { type: string, enum: ['true'] }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
    get:
      operationId: listInvoices
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, CLOSED] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Invoice collection.
          content:
            application/json:
              schema:
                type: object
                required: [items, total, limit, offset]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Invoice' }
                  total: { type: integer, format: int64, minimum: 0, maximum: 9007199254740991 }
                  limit: { type: integer }
                  offset: { type: integer }
        '400': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /invoices/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getInvoice
      responses:
        '200':
          description: Invoice detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /invoices/{id}/close:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
      - name: Idempotency-Key
        in: header
        required: true
        schema: { type: string, minLength: 1, maxLength: 200 }
    post:
      operationId: requestInvoiceClosure
      responses:
        '202':
          description: Durable closure operation accepted.
          headers:
            Idempotent-Replayed:
              description: Present and true when the original operation is replayed.
              schema: { type: string, enum: ['true'] }
          content:
            application/json:
              schema:
                type: object
                required: [operationId, status]
                properties:
                  operationId: { type: string, format: uuid }
                  status: { $ref: '#/components/schemas/OperationStatus' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /invoices/{id}/printable:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getPrintableInvoice
      responses:
        '200':
          description: Closed invoice suitable for a print view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
  /closure-operations/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getClosureOperation
      responses:
        '200':
          description: Current durable operation state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClosureOperation' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '500': { $ref: '#/components/responses/Problem' }
components:
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Problem:
      description: Problem Details error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    CreateInvoiceRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: object
            additionalProperties: false
            required: [productId, quantity]
            properties:
              productId: { type: string, format: uuid }
              quantity: { type: integer, format: int64, minimum: 1, maximum: 9007199254740991 }
    Invoice:
      type: object
      required: [id, number, status, items, version, createdAt, activeClosureOperation]
      properties:
        id: { type: string, format: uuid }
        number: { type: integer, format: int64, minimum: 1, maximum: 9007199254740991 }
        status: { type: string, enum: [OPEN, CLOSED] }
        items:
          type: array
          items:
            type: object
            required: [productId, code, description, quantity]
            properties:
              productId: { type: string, format: uuid }
              code: { type: string }
              description: { type: string }
              quantity: { type: integer, format: int64, minimum: 1, maximum: 9007199254740991 }
        version: { type: integer, format: int64, minimum: 1, maximum: 9007199254740991 }
        createdAt: { type: string, format: date-time }
        closedAt: { type: [string, 'null'], format: date-time }
        activeClosureOperation:
          oneOf:
            - $ref: '#/components/schemas/ActiveClosureOperation'
            - type: 'null'
    ActiveClosureOperation:
      type: object
      required: [operationId, status]
      properties:
        operationId: { type: string, format: uuid }
        status:
          type: string
          enum: [PENDING, PROCESSING, RETRYING]
    ClosureOperation:
      type: object
      required: [id, invoiceId, status, attempts, nextAttemptAt, retryable, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        invoiceId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/OperationStatus' }
        attempts: { type: integer, minimum: 0 }
        nextAttemptAt: { type: string, format: date-time }
        lastError: { type: string }
        retryable: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OperationStatus:
      type: string
      enum: [PENDING, PROCESSING, RETRYING, COMPLETED, FAILED]
    Problem:
      type: object
      required: [type, title, status, code, detail, traceId, retryable, errors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        traceId: { type: string }
        retryable: { type: boolean }
        errors: { type: array, items: { type: object } }